requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.10.0
boto3==1.26.76
botocore==1.29.76
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.28.3
fredapi==0.4.3
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
ipykernel==4.10.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.3.2
llvmlite==0.39.0
lxml==4.6.4
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.1.5
patsy==0.5.2
pmdarima==2.0.3
psycopg2-binary==2.9.5
pymongo==4.3.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
statsmodels==0.13.5
tablib==3.1.0
threadpoolctl==3.0.0
tomli==1.2.2
typing_extensions==4.3.0
urllib3==1.26.7
webencodings==0.5.1
yfinance==0.1.67
zipp==3.8.1
